requirements.txt
# DeployX — Dépendances Python
# pip install -r requirements.txt

# Backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# Frontend
streamlit>=1.33.0
requests>=2.31.0
pandas>=2.2.0
